// apps/server/src/generated/prisma/models/AiUserSerperConfig.ts
export type AiUserSerperConfigModel = runtime.Types.Result.DefaultSelection<Prisma.$AiUserSerperConfigPayload>
// ...
export type $AiUserSerperConfigPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "AiUserSerperConfig"
  objects: {
    user: Prisma.$SystemUserPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    userId: string
    title: string
    apiBase: string
    apiKey: string
    encryptedApiKey: string | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["aiUserSerperConfig"]>
  composites: {}
}